requirements.txt
numpy
scipy
matplotlib
imageio
pandas
scikit-learn
pillow